requirements.txt
python-dotenv
ultralytics
opencv-python
easyocr
flask
numpy
pandas
roboflow